Update wording of LICENSE
Personally I have no issue with the wording of the license. That being said, it is included in Particle's public libraries, and I would love to use this library in a class I am teaching without forking, changing the license wording, and publishing a *nearly* identical library to Particle's database. Without changing the underlying meaning of the license, this wording is simply a little more appropriate for younger students.